Automated_Reproducibility/requirements.txt
django==4.1.7
clearcache-python
python-Levenshtein==0.21.0
GitPython==3.1.29
pymupdf==1.21.0
fuzzywuzzy==0.18.0
pdf2doi==1.5
django-clearcache==1.2.1
requests==2.31.0
ansi2html
beautifulsoup4
clyent==1.2.1
frontend==0.0.3
fuzzywuzzy
ipykernel==6.4.1
matplotlib==3.4.3
numpy==1.23.5
python-time
pytest-shutil
pypdf2==2.0.0
pybase64==1.2.3
pandas==1.3.4
regex==2021.8.3
requests_mock
subprocess.run
statsmodels==0.12.2
scikit-learn==1.3.0
ushlex
zipfile36
jsons
threaded
tools
nbconvert==7.2.9
multiprocess
nbformat==5.1.3
jupyter_client==6.1.12
jax
jaxlib
pyyaml==6.0
seaborn==0.11.2
plotly==5.7.0
kaleido==0.2.1
openpyxl==3.0.9